cleanup(node): move utils functions
