+++ packages/node/src/utils/swc.ts
@@ -1,9 +1,4 @@
-import type { TypeScriptCompilationOptions } from '@nrwl/workspace/src/utilities/typescript';
-import {
-  getNormalizedTsConfig,
-  normalizeOptions,
-} from '@nrwl/workspace/src/utilities/typescript/compilation';
-import { addDependenciesToPackageJson, logger, Tree } from '@nrwl/devkit';
+import { addDependenciesToPackageJson, Tree } from '@nrwl/devkit';
 import {
   swcCoreVersion,
   swcHelpersVersion,
@@ -22,32 +17,6 @@ export function getSwcRuleLoader(): {
   } catch (e) {
     throw new Error(
       '"swc-loader" not installed  in the workspace. Try `npm install --save-dev swc-loader` or `yarn add -D swc-loader`'
-    );
-  }
-}
-
-export function transformTypeScript(tscOptions: TypeScriptCompilationOptions): {
-  success: boolean;
-} {
-  try {
-    const { transformFileSync } = require('@swc/core');
-
-    const normalizedTscOptions = normalizeOptions(tscOptions);
-    const normalizedTsConfig = getNormalizedTsConfig(normalizedTscOptions);
-
-    logger.info(
-      `Compiling TypeScript files for project "${normalizedTscOptions.projectName}"...`
-    );
-    for (const fileName of normalizedTsConfig.fileNames) {
-      transformFileSync(fileName);
-    }
-    logger.info(
-      `Done compiling TypeScript files for project "${normalizedTscOptions.projectName}".`
-    );
-    return { success: true };
-  } catch (e) {
-    throw new Error(
-      '"@swc/core" not installed  in the workspace. Try `npm install --save-dev @swc/core` or `yarn add -D @swc/core`'
     );
   }
 }
